move to scoped package
