+++ README.md
@@ -34,7 +34,7 @@ $ npm install -g @adobe/aio-cli-plugin-config
 $ ./bin/run COMMAND
 running command...
 $ ./bin/run (-v|--version|version)
-@adobe/aio-cli-plugin-config/2.0.0 darwin-x64 node-v8.15.1
+@adobe/aio-cli-plugin-config/2.0.0-rc1 darwin-x64 node-v8.10.0
 $ ./bin/run --help [COMMAND]
 USAGE
   $ ./bin/run COMMAND
@@ -76,7 +76,7 @@ EXAMPLES
   $ aio config:clear
 ```
 
-_See code: [src/commands/config/index.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0/src/commands/config/index.js)_
+_See code: [src/commands/config/index.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0-rc1/src/commands/config/index.js)_
 
 ## `./bin/run config:clear`
 
@@ -92,7 +92,7 @@ OPTIONS
   -l, --local   local config
 ```
 
-_See code: [src/commands/config/clear.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0/src/commands/config/clear.js)_
+_See code: [src/commands/config/clear.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0-rc1/src/commands/config/clear.js)_
 
 ## `./bin/run config:delete KEYS...`
 
@@ -107,11 +107,11 @@ OPTIONS
   -l, --local   local config
 
 ALIASES
-  $ ./bin/run del
-  $ ./bin/run rm
+  $ ./bin/run config:del
+  $ ./bin/run config:rm
 ```
 
-_See code: [src/commands/config/delete.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0/src/commands/config/delete.js)_
+_See code: [src/commands/config/delete.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0-rc1/src/commands/config/delete.js)_
 
 ## `./bin/run config:edit`
 
@@ -126,7 +126,7 @@ OPTIONS
   -l, --local   local config
 ```
 
-_See code: [src/commands/config/edit.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0/src/commands/config/edit.js)_
+_See code: [src/commands/config/edit.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0-rc1/src/commands/config/edit.js)_
 
 ## `./bin/run config:get KEY`
 
@@ -142,7 +142,7 @@ OPTIONS
   -l, --local   local config
 ```
 
-_See code: [src/commands/config/get.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0/src/commands/config/get.js)_
+_See code: [src/commands/config/get.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0-rc1/src/commands/config/get.js)_
 
 ## `./bin/run config:list`
 
@@ -162,7 +162,7 @@ ALIASES
   $ ./bin/run ls
 ```
 
-_See code: [src/commands/config/list.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0/src/commands/config/list.js)_
+_See code: [src/commands/config/list.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0-rc1/src/commands/config/list.js)_
 
 ## `./bin/run config set key 'a value'       # set key to 'a value'`
 
@@ -182,5 +182,5 @@ OPTIONS
   -y, --yaml    value is yaml
 ```
 
-_See code: [src/commands/config/set.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0/src/commands/config/set.js)_
+_See code: [src/commands/config/set.js](https://github.com/adobe/aio-cli-plugin-config/blob/v2.0.0-rc1/src/commands/config/set.js)_
 <!-- commandsstop -->
