+++ src/commands/config/set.js
@@ -15,7 +15,7 @@ const BaseCommand = require('../../base-command')
 const fs = require('fs')
 const yaml = require('js-yaml')
 const hjson = require('hjson')
-const { getPipedData } = require('aio-cli-config')
+const { getPipedData } = require('@adobe/aio-cli-config')
 const { cli } = require('cli-ux')
 const path = require('path')
