+++ src/hooks.js
@@ -1,8 +1,8 @@
-const { dotenv } = require('aio-cli-config')
+const { dotenv } = require('@adobe/aio-cli-config')
 
 /**
  * hoists variables in the ./.env file to process.env
  */
 module.exports = async function() {
-  dotenv(this.debug)
+  dotenv()
 }
