+++ test/__mocks__/@adobe/aio-cli-config.js
